requirements.txt
fastapi
uvicorn
openai
httpx
python-dotenv
openpyxl
deepgram-sdk
whisper
pydub
livekit
soundfile
